Adding weather and location services to backend

+++ InterviewProject.Web/Models/WeatherForecast.cs
@@ -6,7 +6,7 @@ namespace InterviewProject
     {
         public DateTime Date { get; set; }
 
-        public int TemperatureC { get; set; }
+        public double TemperatureC { get; set; }
 
         public int TemperatureF => 32 + (int)(TemperatureC / 0.5556);
